ACL changes

+++ app/code/core/Mage/Adminhtml/controllers/PermissionsController.php
@@ -1,7 +1,8 @@
 <?php
 class Mage_Adminhtml_PermissionsController extends Mage_Adminhtml_Controller_Action
 {
-    public function indexAction() {
+    public function indexAction()
+    {
         $this->loadLayout('baseframe');
         $this->_setActiveMenu('system/acl');
         $this->_addBreadcrumb(__('System'), __('System title'), Mage::getUrl('adminhtml/system'));
@@ -23,7 +24,8 @@ class Mage_Adminhtml_PermissionsController extends Mage_Adminhtml_Controller_Act
         $this->getResponse()->setBody($this->getLayout()->createBlock('adminhtml/permissions_grid_role')->toHtml());
     }
 
-    public function editUserAction() {
+    public function editUserAction()
+    {
         $this->loadLayout('baseframe');
         $this->_addBreadcrumb(__('System'), __('System title'), Mage::getUrl('adminhtml/system'));
         $this->_addBreadcrumb(__('Permission'), __('Permission title'), Mage::getUrl('*/*/index'));
@@ -40,7 +42,8 @@ class Mage_Adminhtml_PermissionsController extends Mage_Adminhtml_Controller_Act
         $this->renderLayout();
     }
 
-    public function editRoleAction() {
+    public function editRoleAction()
+    {
         $this->loadLayout('baseframe');
         $this->_addBreadcrumb(__('System'), __('System title'), Mage::getUrl('adminhtml/system'));
         $this->_addBreadcrumb(__('Permission'), __('Permission title'), Mage::getUrl('*/*/index'));
@@ -59,41 +62,45 @@ class Mage_Adminhtml_PermissionsController extends Mage_Adminhtml_Controller_Act
         $this->renderLayout();
     }
 
-    public function deleteRoleAction() {
+    public function deleteRoleAction()
+    {
     	$rid = $this->getRequest()->getParam('rid', false);
     	Mage::getModel("permissions/roles")->setId($rid)->delete();
 
     	$this->_redirect("adminhtml/permissions");
     }
 
-    public function deleteUserAction() {
+    public function deleteUserAction()
+    {
     	$uid = $this->getRequest()->getParam('uid', false);
     	Mage::getModel("permissions/users")->setId($uid)->delete();
 
     	$this->_redirect("adminhtml/permissions");
     }
 
-    public function saveRoleAction() {
+    public function saveRoleAction()
+    {
     	$rid = $this->getRequest()->getParam('role_id', false);
 
-    	$rid = Mage::getModel("permissions/roles")
+    	$role = Mage::getModel("permissions/roles")
 	    		->setId($rid)
 	    		->setName($this->getRequest()->getParam('role_name', false))
 	    		->setPid($this->getRequest()->getParam('parent_id', false))
+	    		->setRoleType('G')
 	    		->save();
 
     	Mage::getModel("permissions/rules")
-    		->setRoleId($rid)
+    		->setRoleId($role->getId())
     		->setResources($this->getRequest()->getParam('resource', false))
     		->saveRel();
 
 
-    	$rid = explode(",", $rid);
-    	$rid = $rid[0];
-    	$this->_redirect("adminhtml/permissions/editroles/rid/$rid");
+    	$rid = $role->getId();
+    	$this->_redirect("adminhtml/permissions/editrole/rid/$rid");
     }
 
-    public function saveUserAction() {
+    public function saveUserAction()
+    {
     	$uid = $this->getRequest()->getParam('user_id', false);
     	$uid = Mage::getModel("permissions/users")
 	    		->setId($uid)
@@ -112,21 +119,34 @@ class Mage_Adminhtml_PermissionsController extends Mage_Adminhtml_Controller_Act
     	$uid = $uid[0];
     	$this->_redirect("adminhtml/permissions/edituser/uid/$uid");
     }
-    
-    public function deleteuserfromroleAction() {
+
+    public function deleteuserfromroleAction()
+    {
     	Mage::getModel("permissions/users")
-    		->setRoleId($this->getRequest()->getParam('role_id', false))
     		->setUserId($this->getRequest()->getParam('user_id', false))
     		->deleteFromRole();
     	echo json_encode(array('error' => 0, 'error_message' => 'test message'));
     }
-    
-    public function adduser2roleAction() {
-    	Mage::getModel("permissions/users")
-    		->setRoleId($this->getRequest()->getParam('role_id', false))
-    		->setUserId($this->getRequest()->getParam('user_id', false))
-    		->setFirstname($this->getRequest()->getParam('firstname', false))
-    		->add();
-   		echo json_encode(array('error' => 0, 'error_message' => 'test message'));
+
+    public function adduser2roleAction()
+    {
+        if( Mage::getModel("permissions/users")
+        		->setRoleId($this->getRequest()->getParam('role_id', false))
+        		->setUserId($this->getRequest()->getParam('user_id', false))
+                ->roleUserExists() === true ) {
+            echo json_encode(array('error' => 1, 'error_message' => __('This user already added to the role.')));
+        } else {
+        	Mage::getModel("permissions/users")
+        		->setRoleId($this->getRequest()->getParam('role_id', false))
+        		->setUserId($this->getRequest()->getParam('user_id', false))
+        		->setFirstname($this->getRequest()->getParam('firstname', false))
+        		->add();
+       		echo json_encode(array('error' => 0, 'error_message' => 'test message'));
+        }
+    }
+
+    public function editrolegridAction()
+    {
+        $this->getResponse()->setBody($this->getLayout()->createBlock('adminhtml/permissions_role_grid_user')->toHtml());
     }
 }
